Updated integer encoding for support of messages up to 512 bit

+++ RCSwitch.h
@@ -79,7 +79,7 @@ class RCSwitch {
     void switchOff(const char* sHouse, bool bGroup, int nChannel, int nDevice); // NEXA
 
     void sendTriState(const char* sCodeWord);
-    void send(unsigned long long code, unsigned int length);
+    void send(unsigned int code, unsigned int length);
     void send(const char* sCodeWord);
     void sendSpecial(const char* codeWord, int length);
